requirements.txt
# Core Dependencies
numpy>=1.21.0
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.3.0
opencv-python>=4.5.0
scikit-learn>=0.24.0

# Deep Learning & Computer Vision
efficientnet-pytorch>=0.7.0
albumentations>=1.0.0
faiss-cpu>=1.7.0  # Atau faiss-gpu untuk GPU support

# API & Web
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
requests>=2.26.0
aiofiles>=0.7.0

# Monitoring & Deployment
docker>=5.0.0
psutil>=5.8.0
prometheus-client>=0.11.0

# Testing & Development
pytest>=6.2.0
pytest-cov>=2.12.0
black>=21.7b0
flake8>=3.9.0
mypy>=0.910
isort>=5.9.0

# Documentation
mkdocs>=1.2.0
mkdocs-material>=7.2.0

# Utilities
python-dotenv>=0.19.0
tqdm>=4.62.0
pyyaml>=5.4.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Optional Dependencies
# Uncomment jika diperlukan
# tensorboard>=2.6.0
# mlflow>=1.19.0
# wandb>=0.12.0
# ray>=1.5.0
